+++ src/bi/stdio.py
@@ -1,0 +1,8 @@
+def println(*args):
+    """Print args* to the console."""
+    print(*args)
+
+
+def print_(*args):
+    """Print args* to the console."""
+    print(*args, end="")
